+++ crates/miden-testing/src/kernel_tests/tx/test_asset.rs
@@ -21,14 +21,14 @@ fn test_create_fungible_asset_succeeds() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.miden::asset
+        use.miden::faucet
 
         begin
             exec.prologue::prepare_transaction
 
             # create fungible asset
             push.{FUNGIBLE_ASSET_AMOUNT}
-            exec.asset::create_fungible_asset
+            exec.faucet::create_fungible_asset
 
             # truncate the stack
             swapw dropw
@@ -62,14 +62,14 @@ fn test_create_non_fungible_asset_succeeds() -> anyhow::Result<()> {
     let code = format!(
         "
         use.$kernel::prologue
-        use.miden::asset
+        use.miden::faucet
 
         begin
             exec.prologue::prepare_transaction
 
             # push non-fungible asset data hash onto the stack
             push.{non_fungible_asset_data_hash}
-            exec.asset::create_non_fungible_asset
+            exec.faucet::create_non_fungible_asset
 
             # truncate the stack
             swapw dropw